chore: clean up API dependencies
